requirements.txt
flask
flask_mysqldb 
dotenv